+++ pheweb/serve/react/js/components/Chip/ChipTable.tsx
@@ -2,21 +2,15 @@ import React, { useState, useEffect, useMemo, useRef } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import ReactTable from 'react-table-v6'
 import ReactTooltip from 'react-tooltip'
-import { fetchData, setData } from './features/tableSlice'
+import { fetchData, setData, State } from './ChipTableSlice'
 import { chipTableCols } from './tables.js'
 import { mustacheDiv } from "../../common/Utilities";
-import { ChipData } from "./ChipModel";
 import { ConfigurationWindow } from "../Configuration/ConfigurationModel";
 
 const default_banner: string = `
       <p style="paddingBottom: '10px', textDecoration: 'none', color: 'black', width: '200px'">CHIP RESULTS</p>
       `;
 const loading_div = <div>.. . loading . ..</div>
-export type State = {
-  status: string;
-  error: string | null;
-  table: State | null;
-} & Partial<ChipData>;
 
 declare let window: ConfigurationWindow;
 
@@ -32,14 +26,14 @@ export const Table = (props) => {
     
     useEffect(() => {
 	if (data?.status == 'idle') {
-	    dispatch(fetchData(`/api/v1/chip_data`))
+	    dispatch(fetchData(`/api/v1/chip_data`));
 	} else {
 	    const stored = sessionStorage.getItem(`${props.match.params.data}`)
 	    if (stored) {
 		    // console.log('cache hit')
 	    	dispatch(setData(JSON.parse(stored)))
 	    } else if (data?.status != 'loading') {
-		dispatch(fetchData(`/api/v1/chip_data`))
+		dispatch(fetchData(`/api/v1/chip_data`));
 	    }
 	}
     }, [props])
